изменение

+++ calculator.java
@@ -5,10 +5,10 @@ public class calculator {
         Scanner scanner = new Scanner(System.in);
         System.out.println("=== Калькулятор ===");
         System.out.print("Введите первое число: ");
-        int a = scanner.nextInt();
+        double a = scanner.nextInt();
         System.out.print("Введите операцию (+, -, *, /): ");
         String op = scanner.next();
         System.out.print("Введите второе число: ");
-        int b = scanner.nextInt();
+        double b = scanner.nextInt();
     }
 }
